Moved new secure channel implementation into osdp.net lib
This commit introduces SecureChannel2.cs which is intended to eventually
replace the logic in SecureChannel. The new code is based around
IMessageChannel interface which is currently consumed by IncomingMessage
and OSDP.Net.Messages.Reply class.

Whereas original SecureChannel was intended solely for ACU side of the
communications, this new implementation is fully bi-directional and can
support everything original implementation did, but can also represent
the PD side of the secure comms.